+++ pycollocation/__init__.py
@@ -2,16 +2,15 @@
 Objects imported here will live in the `pycollocation` namespace
 
 """
-__all__ = ["SymbolicBoundaryValueProblem", "OrthogonalPolynomialSolver", "Visualizer"]
+__all__ = ["TwoPointBVP", "TwoPointBVPLike", "SymbolicTwoPointBVPLike",
+           "SymbolicTwoPointBVP", "OrthogonalPolynomialSolver", "Visualizer"]
 
-
-from . import boundary_value_problems
-from . import differential_equations
+from . import bvp
 from . import models
 from . import orthogonal_polynomials
 from . import visualizers
 
-from . boundary_value_problems import SymbolicBoundaryValueProblem
+from . bvp import *
 from . orthogonal_polynomials import OrthogonalPolynomialSolver
 from . visualizers import Visualizer
