+++ pycollocation/models.py
@@ -1,4 +1,14 @@
+"""
+Classes for representing differential and difference equations.
+
+@author : David R. Pugh
+
+"""
 import collections
+
+import sympy as sym
+
+import symbolics
 
 
 class ModelLike(object):
@@ -9,28 +19,40 @@ class ModelLike(object):
         Model dependent variables.
 
         :getter: Return the model dependent variables.
+        :setter: Set new model dependent variables.
         :type: list
 
         """
         return self._dependent_vars
 
+    @dependent_vars.setter
+    def dependent_vars(self, variables):
+        """Set new model dependent variables."""
+        self._dependent_vars = self._validate_variables(variables)
+
     @property
     def independent_var(self):
         """
-        Symbolic variable representing the independent variable.
+        Model independent variable.
 
-        :getter: Return the symbol representing the independent variable.
-        :type: sympy.Symbol
+        :getter: Return the model independent variable.
+        :setter: Set new model independent variable.
+        :type: string
 
         """
         return self._independent_var
+
+    @independent_var.setter
+    def independent_var(self, variable):
+        """Set new model independent variable."""
+        self._independent_var = self._validate_variable(variable)
 
     @property
     def params(self):
         """
         Dictionary of model parameters.
 
-        :getter: Return the current parameter dictionary.
+        :getter: Return the model parameters as an ordered dictionary.
         :setter: Set a new parameter dictionary.
         :type: dict
 
@@ -46,14 +68,19 @@ class ModelLike(object):
     @property
     def rhs(self):
         """
-        Symbolic representation of the right-hand side of a system of
-        differential/difference equations.
+        Right-hand side of the system of differential/difference equations.
 
         :getter: Return the right-hand side of the system of equations.
+        :setter: Set new value for the right-hand side of the system of equations.
         :type: dict
 
         """
         return self._rhs
+
+    @rhs.setter
+    def rhs(self, rhs):
+        """Set new value for the right-hand side of the system of equations."""
+        self._rhs = self._validate_rhs(rhs)
 
     @staticmethod
     def _order_params(params):
@@ -68,3 +95,92 @@ class ModelLike(object):
             raise AttributeError(mesg.format(value.__class__))
         else:
             return value
+
+    def _validate_rhs(self, rhs):
+        """Validate the rhs attribute."""
+        if not isinstance(rhs, dict):
+            mesg = "Attribute `rhs` must be of type `dict` not {}"
+            raise AttributeError(mesg.format(rhs.__class__))
+        elif not (len(rhs) == len(self.dependent_vars)):
+            mesg = "Number of equations must equal number of dependent vars."
+            raise ValueError(mesg)
+        else:
+            return rhs
+
+    @staticmethod
+    def _validate_variable(symbol):
+        """Validate the independent_var attribute."""
+        if not isinstance(symbol, str):
+            mesg = "Attribute must be of type `string` not {}"
+            raise AttributeError(mesg.format(symbol.__class__))
+        else:
+            return symbol
+
+    @classmethod
+    def _validate_variables(cls, symbols):
+        """Validate the dependent_vars attribute."""
+        return [cls._validate_variable(symbol) for symbol in symbols]
+
+
+class SymbolicModelLike(symbolics.SymbolicLike, ModelLike):
+
+    _cached_rhs_functions = {}  # not sure if this is good practice!
+
+    __symbolic_jacobian = None
+
+    @property
+    def _symbolic_args(self):
+        """List of symbolic arguments used to lambdify expressions."""
+        return self._symbolic_vars + self._symbolic_params.values()
+
+    @property
+    def _symbolic_jacobian(self):
+        """Symbolic Jacobian matrix of partial derivatives."""
+        if self.__symbolic_jacobian is None:
+            args = self.dependent_vars
+            self.__symbolic_jacobian = self._symbolic_system.jacobian(args)
+        return self.__symbolic_jacobian
+
+    @property
+    def _symbolic_params(self):
+        """List of symbolic model parameters."""
+        return collections.OrderedDict((k, sym.symbols(k)) for (k, v) in self.params.items())
+
+    @property
+    def _symbolic_system(self):
+        """Represents rhs as a symbolic matrix."""
+        return sym.Matrix([self.rhs[var] for var in self.dependent_vars])
+
+    @property
+    def _symbolic_vars(self):
+        """List of symbolic model variables."""
+        return sym.symbols([self.independent_var] + self.dependent_vars)
+
+    def _clear_cache(self):
+        """Clear cached symbolic Jacobian."""
+        self.__symbolic_jacobian = None
+
+    def _rhs_functions(self, var):
+        """Cache lamdified rhs functions for numerical evaluation."""
+        if self._cached_rhs_functions.get(var) is None:
+            eqn = self.rhs[var]
+            self._cached_rhs_functions[var] = self._lambdify_factory(eqn)
+        return self._cached_rhs_functions[var]
+
+    @staticmethod
+    def _validate_expression(expression):
+        """Validates a symbolic expression."""
+        if not isinstance(expression, sym.Basic):
+            mesg = "Attribute must be of type `sympy.Basic` not {}"
+            raise AttributeError(mesg.format(expression.__class__))
+        else:
+            return expression
+
+    def _validate_rhs(self, rhs):
+        """Validate a the rhs attribute."""
+        super(SymbolicModelLike, self)._validate_rhs(rhs)
+        exprs = {}
+        for var, expr in rhs.items():
+            exprs[var] = self._validate_expression(expr)
+        else:
+            return exprs
