--- pycollocation/tests/test_symbolics.py
@@ -1,10 +1,0 @@
-import nose
-
-from .. import symbolics
-
-
-def test_symbolic_base():
-    """SymbolicsBase._symbolic_args attribute is not implemented."""
-    with nose.tools.assert_raises(NotImplementedError):
-        base = symbolics.SymbolicBase()
-        base._symbolic_args
